Initial upload
Bittrex_Trend_Alert pulls the latest trade for BTC, ETH, and OMG and compares the value to CryptoDB.json (see Bittrex_BD_Populate). The script does some simple statistical analysis and linear regression to detect if a sizable change has occured. If the last trade is within the normal trading ranges than the script stores the statistical data in CryptoStats.json (for a spike alert algorithim). If the last trade is outside of the norm, then an alert is sent to a slack channel with a webhook.