Updated Bittrex_Trend_Update
The threshold value previously used a linear curve fit based upon magnitude. The rise of BitCoin over $10,000 resulted in a step drop of the threshold values which would leave to odd behavior on either side of the $10,000 mark. I took a look at the data and removed the step function to the threshold value. The threshold value should now be linear when graphed against the current price in a log-log or will be a power function if graphed in log(price) vs threshold value. Test values showed that this behavior should hold however I will monitor the behavior and tweak as needed.